+++ CHANGELOG.md
@@ -8,3 +8,4 @@
  - Add "Has cast spell" requirement
  - Add "Cast spell" objective
  - Add "Has Wand" requirement
+ - Add "active wand" and "active spell" requirements. Not sure if these are useful.